become more of a rust wannabe
- add f32 and f64 types
- rebrand "int.h" to "types.h"
- add progress arg to git submodule commands
- update some cmake logs

+++ lib/nav.c
@@ -4,22 +4,21 @@
  */
 
 #include <math.h>
-#include "platform/int.h"
 
 #include "nav.h"
 
 #define EARTH_RADIUS_KM 6371                    // Earth's radius in kilometers
 #define EARTH_RADIUS_M (EARTH_RADIUS_KM * 1000) // Earth's radius in meters
 
-double calculate_bearing(long double latA, long double lonA, long double latB, long double lonB) {
-    long double thetaA = radians(latA);
-    long double thetaB = radians(latB);
-    long double deltaL = radians(lonB - lonA);
+f64 calculate_bearing(f64 latA, f64 lonA, f64 latB, f64 lonB) {
+    f64 thetaA = radians(latA);
+    f64 thetaB = radians(latB);
+    f64 deltaL = radians(lonB - lonA);
 
-    double X = cos((double)thetaB) * sin((double)deltaL);
-    double Y = cos((double)thetaA) * sin((double)thetaB) - sin((double)thetaA) * cos((double)thetaB) * cos((double)deltaL);
+    f64 X = cos(thetaB) * sin(deltaL);
+    f64 Y = cos(thetaA) * sin(thetaB) - sin(thetaA) * cos(thetaB) * cos(deltaL);
 
-    double bearing = atan2(X, Y) * 180.0 / M_PI;
+    f64 bearing = atan2(X, Y) * 180.0 / M_PI;
     // Map to a heading instead of a degree value
     if (bearing <= 0) {
         bearing += 360;
@@ -27,15 +26,15 @@ double calculate_bearing(long double latA, long double lonA, long double latB, l
     return bearing;
 }
 
-double calculate_distance(long double latA, long double lonA, long double latB, long double lonB) {
-    long double thetaA = radians(latA);
-    long double thetaB = radians(latB);
-    long double deltaT = radians(latB - latA);
-    long double deltaL = radians(lonB - lonA);
+f64 calculate_distance(f64 latA, f64 lonA, f64 latB, f64 lonB) {
+    f64 thetaA = radians(latA);
+    f64 thetaB = radians(latB);
+    f64 deltaT = radians(latB - latA);
+    f64 deltaL = radians(lonB - lonA);
 
-    double a = sin((double)deltaT / 2) * sin((double)deltaT / 2) +
-               cos((double)thetaA) * cos((double)thetaB) * sin((double)deltaL / 2) * sin((double)deltaL / 2);
-    double c = 2 * atan2(sqrt(a), sqrt(1 - a));
+    f64 a = sin(deltaT / 2) * sin(deltaT / 2) +
+               cos(thetaA) * cos(thetaB) * sin(deltaL / 2) * sin(deltaL / 2);
+    f64 c = 2 * atan2(sqrt(a), sqrt(1 - a));
 
     return EARTH_RADIUS_M * c;
 }
